src/requirements.txt
beautifulsoup4==4.11.1
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
PyQt5==5.15.7
requests==2.28.1
selenium==4.7.0
webdriver_manager==3.8.5
